Fixes #11:  make ToServerAddrs work with IPv6 addresses
Previously ToServerAddrs considered colon in IPv6 six address as a port
separator and did not add the port when it was needed. Now we use a bit
more heuristics to determine f we have an IPv6 address. The best
solution would be to match the string with a regexp but I don't want to
introduce a regex crate dependency just for that.