Fix "connect_ip" config behaviour
If "connect_ip" config param were set to "true" then both the UDP
socket has been "connected" and send_to is used to send the packet.
This caused problem when the remote address was a DNS name which
could yield to multiple IP addresses, as in those cases connect and
send_to might got different addresses.